# platform/api-contracts/inference-gateway.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReceiptRequest": {
        "additionalProperties": false,
        "description": "A single agent action reported by a workspace for the audit chain.\n\nEvery free-text field is bounded and redacted before it is chained. A receipt is\npermanent, tamper-evident, and read by auditors, which makes it the last place a\ncredential pasted into a command line should be allowed to come to rest. A producer\nthat needs to commit to detail it cannot publish sends ``detail_sha256`` instead.",
        "properties": {
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "correlation_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Request Id"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "detail_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail Sha256"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "sandbox_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Id"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "tool": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool"
          }
        },
        "required": [
          "action_type",
          "decision"
        ],
        "title": "ReceiptRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "API key header accepted by gateway middleware when API key authentication is enabled.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "description": "Bearer token accepted by gateway middleware when API key authentication is enabled.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "OpenAI-compatible private inference gateway with sandbox traceability, admission controls, budget enforcement, redacted audit events, and runtime routing to Ollama or vLLM.",
    "title": "Private AI Platform Kit Inference Gateway",
    "version": "0.29.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/models": {
      "get": {
        "operationId": "listModels",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Listmodels",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "List approved private models",
        "tags": [
          "inference"
        ]
      }
    },
...
    "/v1/receipts": {
      "post": {
        "description": "Agent-action receipt intake (ADR 0014). Records what a workspace did beyond calling a model (denied egress, tool execution, file writes, credential requests) on the SAME tamper-evident hash chain as model calls, so 'make audit-verify' checks them together. The action_type vocabulary is closed; an unknown value is rejected. This endpoint RECORDS claims and enforces nothing: a receipt reporting a denial is evidence that some other control denied the action, never the control itself, and submitting one permits nothing. The sandbox on the receipt is the caller's bound sandbox, so a workspace can only add to its own history. Off by default (AGENT_RECEIPTS_ENABLED).",
        "operationId": "createAgentReceipt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Createagentreceipt",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Record an agent action on the audit chain",
        "tags": [
          "sandbox"
        ]
      }
    },
...
    "/v1/sandbox/budget": {
      "get": {
        "operationId": "getSandboxBudget",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Getsandboxbudget",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Get sandbox budget usage",
        "tags": [
          "sandbox"
        ]
      }
    },
    "/v1/usage": {
      "get": {
        "operationId": "getSandboxUsage",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Getsandboxusage",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Get sandbox usage and estimated cost",
        "tags": [
          "sandbox"
        ]
      }
    }
  },
...
}